refactor: adjustement security and payment

+++ src/migrations/1775319070624-AddAcceptedTermsToUser.ts
@@ -4,11 +4,21 @@ export class AddAcceptedTermsToUser1775319070624 implements MigrationInterface {
     name = 'AddAcceptedTermsToUser1775319070624'
 
     public async up(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.query(`ALTER TABLE "users" ADD "accepted_terms" boolean NOT NULL DEFAULT false`);
+        const table = await queryRunner.getTable("users");
+        const hasColumn = table?.findColumnByName("accepted_terms");
+        
+        if (!hasColumn) {
+            await queryRunner.query(`ALTER TABLE "users" ADD "accepted_terms" boolean NOT NULL DEFAULT false`);
+        }
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.query(`ALTER TABLE "users" DROP COLUMN "accepted_terms"`);
+        const table = await queryRunner.getTable("users");
+        const hasColumn = table?.findColumnByName("accepted_terms");
+
+        if (hasColumn) {
+            await queryRunner.query(`ALTER TABLE "users" DROP COLUMN "accepted_terms"`);
+        }
     }
 
 }
